+++ verticapy/machine_learning/model_selection/model_selection.py
@@ -42,7 +42,7 @@ from verticapy.plotting._matplotlib import (
     plot_stepwise_ml,
     plot_importance,
 )
-from verticapy.plotting._matplotlib.core import updated_dict
+from verticapy.plotting._matplotlib.base import updated_dict
 from verticapy.sql._utils._format import quote_ident, schema_relation
 from verticapy.machine_learning._utils import reverse_score
 from verticapy._config.config import OPTIONS
