+++ verticapy/plotting/_highcharts/__init__.py
@@ -14,4 +14,4 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-from verticapy.plotting._highcharts.highchart import *
+from verticapy.plotting._highcharts.base import *
