+++ verticapy/plotting/_highcharts/base.py
