+++ verticapy/plotting/_matplotlib/__init__.py
@@ -20,7 +20,7 @@ from verticapy.plotting._matplotlib.animated import (
     animated_ts_plot,
 )
 from verticapy.plotting._matplotlib.bar import bar, bar2D, hist, hist2D, multiple_hist
-from verticapy.plotting._matplotlib.core import compute_plot_variables
+from verticapy.plotting._matplotlib.base import compute_plot_variables
 from verticapy.plotting._matplotlib.boxplot import boxplot, boxplot2D
 from verticapy.plotting._matplotlib.heatmap import (
     cmatrix,
