+++ verticapy/plotting/_matplotlib/animated.py
@@ -26,7 +26,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 # VerticaPy Modules
-from verticapy.plotting._matplotlib.core import updated_dict
+from verticapy.plotting._matplotlib.base import updated_dict
 from verticapy._config.config import ISNOTEBOOK
 from verticapy._utils._sql import _executeSQL
 from verticapy.plotting._colors import gen_cmap, gen_colors
