+++ verticapy/plotting/_matplotlib/bar.py
@@ -22,10 +22,10 @@ import matplotlib.ticker as mticker
 import matplotlib.pyplot as plt
 
 # VerticaPy Modules
-from verticapy.plotting._matplotlib.core import updated_dict
+from verticapy.plotting._matplotlib.base import updated_dict
 from verticapy._config.config import ISNOTEBOOK
 from verticapy.errors import ParameterError
-from verticapy.plotting._matplotlib.core import compute_plot_variables
+from verticapy.plotting._matplotlib.base import compute_plot_variables
 from verticapy.plotting._colors import gen_colors
