+++ verticapy/plotting/_matplotlib/heatmap.py
@@ -27,7 +27,7 @@ import numpy as np
 
 # VerticaPy Modules
 from verticapy.core.tablesample import tablesample
-from verticapy.plotting._matplotlib.core import updated_dict
+from verticapy.plotting._matplotlib.base import updated_dict
 from verticapy._utils._cast import to_varchar
 from verticapy._config.config import ISNOTEBOOK
 from verticapy.sql.read import to_tablesample
