+++ verticapy/plotting/_matplotlib/mlplot.py
@@ -35,7 +35,7 @@ from verticapy._utils._sql import _executeSQL
 from verticapy.errors import ParameterError
 from verticapy.plotting._colors import gen_colors, get_color
 from verticapy.sql._utils._format import quote_ident
-from verticapy.plotting._matplotlib.core import updated_dict
+from verticapy.plotting._matplotlib.base import updated_dict
 
 
 def logit_plot(
