+++ verticapy/plotting/_matplotlib/pie.py
@@ -22,9 +22,9 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 # VerticaPy Modules
-from verticapy.plotting._matplotlib.core import updated_dict
+from verticapy.plotting._matplotlib.base import updated_dict
 from verticapy._config.config import ISNOTEBOOK
-from verticapy.plotting._matplotlib.core import compute_plot_variables
+from verticapy.plotting._matplotlib.base import compute_plot_variables
 from verticapy.plotting._colors import gen_colors
