+++ verticapy/plotting/_matplotlib/timeseries.py
@@ -21,7 +21,7 @@ import warnings
 import matplotlib.pyplot as plt
 
 # VerticaPy Modules
-from verticapy.plotting._matplotlib.core import updated_dict
+from verticapy.plotting._matplotlib.base import updated_dict
 from verticapy._config.config import ISNOTEBOOK
 from verticapy._utils._sql import _executeSQL
 from verticapy.plotting._colors import get_color, gen_colors
